+++ src/components/DeviceSelectionDialog/DeviceSelectionDialog.tsx
@@ -12,10 +12,19 @@ import {
   Theme,
   DialogTitle,
   Hidden,
+  FormControlLabel,
+  Switch,
+  Tooltip,
 } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import VideoInputList from './VideoInputList/VideoInputList';
 import MaxGalleryViewParticipants from './MaxGalleryViewParticipants/MaxGalleryViewParticipants';
+import { useKrispToggle } from '../../hooks/useKrispToggle/useKrispToggle';
+import SmallCheckIcon from '../../icons/SmallCheckIcon';
+import InfoIconOutlined from '../../icons/InfoIconOutlined';
+import KrispLogo from '../../icons/KrispLogo';
+import { useAppState } from '../../state';
+import useVideoContext from '../../hooks/useVideoContext/useVideoContext';
 
 const useStyles = makeStyles((theme: Theme) => ({
   container: {
@@ -46,9 +55,28 @@ const useStyles = makeStyles((theme: Theme) => ({
       margin: '1em 0 2em 0',
     },
   },
+  noiseCancellationContainer: {
+    display: 'flex',
+    justifyContent: 'space-between',
+  },
+  krispContainer: {
+    display: 'flex',
+    alignItems: 'center',
+    '& svg': {
+      '&:not(:last-child)': {
+        margin: '0 0.3em',
+      },
+    },
+  },
+  krispInfoText: {
+    margin: '0 0 1.5em 0.5em',
+  },
 }));
 
 export default function DeviceSelectionDialog({ open, onClose }: { open: boolean; onClose: () => void }) {
+  const { isAcquiringLocalTracks } = useVideoContext();
+  const { isKrispEnabled, isKrispInstalled } = useAppState();
+  const { toggleKrisp } = useKrispToggle();
   const classes = useStyles();
 
   return (
@@ -67,6 +95,45 @@ export default function DeviceSelectionDialog({ open, onClose }: { open: boolean
           <Typography variant="h6" className={classes.headline}>
             Audio
           </Typography>
+
+          {isKrispInstalled && (
+            <div className={classes.noiseCancellationContainer}>
+              <div className={classes.krispContainer}>
+                <Typography variant="subtitle2">Noise Cancellation powered by </Typography>
+                <KrispLogo />
+                <Tooltip
+                  title="Suppress background noise from your microphone"
+                  interactive
+                  leaveDelay={250}
+                  leaveTouchDelay={15000}
+                  enterTouchDelay={0}
+                >
+                  <div>
+                    <InfoIconOutlined />
+                  </div>
+                </Tooltip>
+              </div>
+              <FormControlLabel
+                control={
+                  <Switch
+                    checked={!!isKrispEnabled}
+                    checkedIcon={<SmallCheckIcon />}
+                    disableRipple={true}
+                    onClick={toggleKrisp}
+                  />
+                }
+                label={isKrispEnabled ? 'Enabled' : 'Disabled'}
+                style={{ marginRight: 0 }}
+                disabled={isAcquiringLocalTracks}
+              />
+            </div>
+          )}
+          {isKrispInstalled && (
+            <Typography variant="body1" color="textSecondary" className={classes.krispInfoText}>
+              Suppress background noise from your microphone.
+            </Typography>
+          )}
+
           <AudioInputList />
         </div>
         <div className={classes.listSection}>